Remove hooks/utils
Given that pre-commit is started from a different directory than
the pre-commit-hooks directory, it takes some legwork to change
and then change back to user's code directory.

Instead of having shared utils, each script should take of itself
and not use shared functions. This results in duplicate code,
but it's only 5 lines each.