Tests work now for both pip command and command class
When you install this library with pip, it will come with commands like

clang-format-hook
clang-tidy-hook
oclint-hook

Essentially your favorite C Linter with -hook at the end. These will be
accessible via the command line and are tested here.

The other tests are done on the python classes themselves, which keep
track of stdout/stderr/returncode for this to work.